refactor: metro controller
