+++ backend-nest/src/app.service.ts
@@ -1,9 +1,6 @@
 import { Injectable } from "@nestjs/common";
 
-@Injectable()
-export class AppService {
-    getHello(): string {
-        return `
+export const WELCOME_MESSAGE = `
             <h1>
                 Hello, fellow developer👋
             </h1>
@@ -12,5 +9,10 @@ export class AppService {
                 You can find the API documentation <a href="https://metro-now.vercel.app/docs/rest-api">here</a>.
             </p>
         `;
+
+@Injectable()
+export class AppService {
+    getHello(): string {
+        return WELCOME_MESSAGE;
     }
 }
